pyproject.toml
[project]
name = "trajectly-survival-arena"
version = "0.1.0"
description = "GitHub-native agent evolution game powered by Trajectly"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "trajectly==0.4.2",
  "Pillow>=10.0.0,<12.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

